+++ src/main/java/com/buuz135/sushigocrafting/datagen/SushiItemTagsProvider.java
@@ -40,11 +40,11 @@ public class SushiItemTagsProvider extends ItemTagsProvider {
         this.copy(BlockTags.SAPLINGS, ItemTags.SAPLINGS);
         for (CustomCropBlock block : new CustomCropBlock[]{SushiContent.Blocks.RICE_CROP.get(), SushiContent.Blocks.CUCUMBER_CROP.get(), SushiContent.Blocks.SOY_CROP.get(), SushiContent.Blocks.WASABI_CROP.get(), SushiContent.Blocks.SESAME_CROP.get()}) {
             tag(Tags.Items.SEEDS).add(block.getBaseSeedId().asItem());
-            tag(ItemTags.create(ResourceLocation.fromNamespaceAndPath("c", "seeds/" + BuiltInRegistries.ITEM.getKey(block.getBaseSeedId().asItem()).getPath()))).add(block.getBaseSeedId().asItem());
+            tag(ItemTags.create(ResourceLocation.fromNamespaceAndPath("c", "seeds/" + getNameFor(BuiltInRegistries.ITEM.getKey(block.getBaseSeedId().asItem()).getPath())))).add(block.getBaseSeedId().asItem());
         }
         for (Item item : new Item[]{SushiContent.Items.RICE.get(), SushiContent.Items.CUCUMBER.get(), SushiContent.Items.SOY_BEAN.get(), SushiContent.Items.WASABI_ROOT.get(), SushiContent.Items.SESAME_SEED.get()}) {
             tag(Tags.Items.CROPS).add(item);
-            tag(ItemTags.create(ResourceLocation.fromNamespaceAndPath("c", "crops/" + BuiltInRegistries.ITEM.getKey(item).getPath()))).add(item.asItem());
+            tag(ItemTags.create(ResourceLocation.fromNamespaceAndPath("c", "crops/" + getNameFor(BuiltInRegistries.ITEM.getKey(item).getPath())))).add(item.asItem());
         }
         tag(ItemTags.FISHES).add(SushiContent.Items.RAW_TUNA.get()).add(SushiContent.Items.SHRIMP.get());
         tag(ItemTags.create(ResourceLocation.fromNamespaceAndPath("c", "raw_fishes"))).add(Items.SALMON).add(SushiContent.Items.SHRIMP.get()).add(SushiContent.Items.RAW_TUNA.get());
@@ -63,5 +63,10 @@ public class SushiItemTagsProvider extends ItemTagsProvider {
                 }
             }
         }
+    }
+
+    public String getNameFor(String name){
+        if (name.equals("soy_bean")) return "soybean";
+        return name.replaceAll("_seeds", "");
     }
 }