+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Added German Translation by Stein-N
 * Update spanish translation to reflect all the keys in the en_us.json by Witixin1512
 * PistonCrafting with Tags by TheonlyTazz
+* Changed soy_bean tag to soybean and removed "_seeds" from seed tags, closes #54 closes #52 closes #50